# contracts/fastapi-openapi-v1.json
{
  "components": {
    "schemas": {
...
      "CallAuctionMarketSeriesSnapshotQuery": {
        "additionalProperties": false,
        "properties": {
          "codes": {
            "description": "不含交易所前缀的六位股票代码集合，最多五百个。",
            "items": {
              "pattern": "^[0-9]{6}$",
              "type": "string"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Codes",
            "type": "array"
          },
          "trade_date": {
            "description": "交易日。",
            "format": "date",
            "title": "Trade Date",
            "type": "string"
          }
        },
        "required": [
          "trade_date",
          "codes"
        ],
        "title": "CallAuctionMarketSeriesSnapshotQuery",
        "type": "object"
      },
...
      "CallAuctionMarketSnapshotQuery": {
        "additionalProperties": false,
        "properties": {
          "codes": {
            "description": "不含交易所前缀的六位股票代码集合，最多五百个。",
            "items": {
              "pattern": "^[0-9]{6}$",
              "type": "string"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Codes",
            "type": "array"
          },
          "trade_date": {
            "description": "交易日。",
            "format": "date",
            "title": "Trade Date",
            "type": "string"
          }
        },
        "required": [
          "trade_date",
          "codes"
        ],
        "title": "CallAuctionMarketSnapshotQuery",
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "不含交易所前缀的六位股票代码。",
            "title": "Code",
            "type": "string"
          },
          "message": {
            "description": "面向调用方的错误说明。",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "错误详情。"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "提供有界、只读且可追溯的证券、日线、涨停池、集合竞价及客观衍生数据查询。除健康检查外，业务接口均需通过 X-API-Key 鉴权。",
    "summary": "面向外部消费者的只读 A 股市场数据接口",
    "title": "股票市场数据中心 API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/board-indexes/883423/bias": {
      "get": {
        "description": "只读取数据库中最新已持久化的 THS:883423 日线。返回当前五日简单移动平均乖离率、相对上一有效交易日的变化方向，以及近三十个交易日有效样本的最高和最低乖离率。当天尚未入库时返回数据库最近交易日并明确 trade_date；历史少于34条时返回404。本接口不接收参数，不访问行情提供方，也不写入数据。",
        "operationId": "board_index_bias_latest_api_v1_board_indexes_883423_bias_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardIndexBiasResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "查询同花顺 883423 板块最新五日线乖离指标",
        "tags": [
          "市场数据"
        ]
      }
    },
    "/api/v1/call-auction-indicative-details": {
      "get": {
        "description": "接收一个六位沪深股票代码，优先读取上海时区当日最新数据库快照。仅在数据库明确无数据时，有界访问东方财富，获取 09:15:00 至 09:25:59 的虚拟匹配或参考价格及展示匹配量。实时结果在保存不可变原始数据后立即返回，数据库登记异步执行。该数据不是交易所成交明细或逐笔委托，来源展示分类也不代表成交方向。结果按观察时间和来源序号正序排列，时间格式为上海时区 YYYY-MM-DD HH:mm:ss。",
        "operationId": "auction_indicative_details_api_v1_call_auction_indicative_details_get",
        "parameters": [
          {
            "description": "不含交易所前缀的六位股票代码。",
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "pattern": "^[0-9]{6}$",
              "title": "Code",
              "type": "string"
            }
          },
          {
            "description": "从结果集起点跳过的记录数。",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 5000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "最多返回的记录数。",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionIndicativeDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "查询当日集合竞价虚拟匹配明细",
        "tags": [
          "市场数据"
        ]
      }
    },
    "/api/v1/call-auction-market-series-snapshots/query": {
      "post": {
        "description": "返回指定交易日最新成功采集会话的全部轮次；仅在没有成功会话时使用最新部分成功会话。轮次按计划采集时间正序排列，并分别报告未命中的六位代码。不同会话或日期的数据不会合并或替代。",
        "operationId": "call_auction_market_series_snapshots_api_v1_call_auction_market_series_snapshots_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallAuctionMarketSeriesSnapshotQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallAuctionMarketSeriesSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "批量查询开盘集合竞价序列快照",
        "tags": [
          "市场数据"
        ]
      }
    },
    "/api/v1/call-auction-market-snapshots/query": {
      "post": {
        "description": "返回指定交易日最新成功采集批次的数据；仅在没有成功批次时使用最新部分成功批次。同一个六位代码可能同时匹配沪市和深市证券，未命中的代码会明确列出。不使用日期或其他批次回退。",
        "operationId": "call_auction_market_snapshots_api_v1_call_auction_market_snapshots_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallAuctionMarketSnapshotQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallAuctionMarketSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "批量查询开盘集合竞价市场快照",
        "tags": [
          "市场数据"
        ]
      }
    },
    "/api/v1/call-auction-one-price-limits": {
      "get": {
        "description": "选取一批已存储的沪深市场 09:25:30 快照，在读取时按已接受的主板百分之十涨跌停规则计算结果。不依赖夜间涨跌停批次，不访问行情提供方，不使用更晚的日线，也不写入数据。",
        "operationId": "auction_one_price_limits_api_v1_call_auction_one_price_limits_get",
        "parameters": [
          {
            "description": "交易日。",
            "in": "query",
            "name": "trade_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trade Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionOnePriceLimitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "实时计算 09:25:30 主板一字涨跌停列表",
        "tags": [
          "市场数据"
        ]
      }
    },
...
    "/api/v1/close-price-new-highs-120d": {
      "get": {
        "description": "返回最新交易日收盘价严格高于此前一百一十九个交易日最高收盘价的股票。",
        "operationId": "close_price_new_highs_120d_api_v1_close_price_new_highs_120d_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosePriceNewHighs120dResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "查询沪深两市收盘价创一百二十日新高的股票",
        "tags": [
          "市场数据"
        ]
      }
    },
...
  },
...
}